Split SQL injection analyzer into separate EF and ADO injection analyzers.
